+++ tests/mindee_http/test_error.py
@@ -6,8 +6,8 @@ import pytest
 from mindee import Client, product
 from mindee.input.sources import PathInput
 from mindee.mindee_http.error import (
-    MindeeHTTPClientException,
-    MindeeHTTPServerException,
+    MindeeHTTPClientError,
+    MindeeHTTPServerError,
     handle_error,
 )
 from tests.test_inputs import FILE_TYPES_DIR
@@ -36,24 +36,24 @@ def dummy_file(monkeypatch) -> PathInput:
 
 
 def test_http_client_error(dummy_client: Client, dummy_file: PathInput):
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         dummy_client.parse(product.InvoiceV4, dummy_file)
 
 
 def test_http_enqueue_client_error(dummy_client: Client, dummy_file: PathInput):
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         dummy_client.enqueue(product.InvoiceV4, dummy_file)
 
 
 def test_http_parse_client_error(dummy_client: Client, dummy_file: PathInput):
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         dummy_client.parse_queued(product.InvoiceV4, "dummy-queue-id")
 
 
 def test_http_enqueue_and_parse_client_error(
     dummy_client: Client, dummy_file: PathInput
 ):
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         dummy_client.enqueue_and_parse(product.InvoiceV4, dummy_file)
 
 
@@ -61,7 +61,7 @@ def test_http_400_error():
     error_ref = open(ERROR_DATA_DIR / "error_400_no_details.json")
     error_obj = json.load(error_ref)
     error_400 = handle_error("dummy-url", error_obj, 400)
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         raise error_400
     assert error_400.status_code == 400
     assert error_400.api_code == "SomeCode"
@@ -73,7 +73,7 @@ def test_http_401_error():
     error_ref = open(ERROR_DATA_DIR / "error_401_invalid_token.json")
     error_obj = json.load(error_ref)
     error_401 = handle_error("dummy-url", error_obj, 401)
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         raise error_401
     assert error_401.status_code == 401
     assert error_401.api_code == "Unauthorized"
@@ -85,7 +85,7 @@ def test_http_429_error():
     error_ref = open(ERROR_DATA_DIR / "error_429_too_many_requests.json")
     error_obj = json.load(error_ref)
     error_429 = handle_error("dummy-url", error_obj, 429)
-    with pytest.raises(MindeeHTTPClientException):
+    with pytest.raises(MindeeHTTPClientError):
         raise error_429
     assert error_429.status_code == 429
     assert error_429.api_code == "TooManyRequests"
@@ -97,7 +97,7 @@ def test_http_500_error():
     error_ref = open(ERROR_DATA_DIR / "error_500_inference_fail.json")
     error_obj = json.load(error_ref)
     error_500 = handle_error("dummy-url", error_obj, 500)
-    with pytest.raises(MindeeHTTPServerException):
+    with pytest.raises(MindeeHTTPServerError):
         raise error_500
     assert error_500.status_code == 500
     assert error_500.api_code == "failure"
@@ -108,7 +108,7 @@ def test_http_500_error():
 def test_http_500_html_error():
     error_ref_contents = open(ERROR_DATA_DIR / "error_50x.html").read()
     error_500 = handle_error("dummy-url", error_ref_contents, 500)
-    with pytest.raises(MindeeHTTPServerException):
+    with pytest.raises(MindeeHTTPServerError):
         raise error_500
     assert error_500.status_code == 500
     assert error_500.api_code == "UnknownError"
