+++ tests/test_client.py
@@ -5,7 +5,7 @@ import pytest
 from mindee import Client, PageOptions, product
 from mindee.error.mindee_error import MindeeClientError
 from mindee.input.sources import LocalInputSource
-from mindee.mindee_http.error import MindeeHTTPException
+from mindee.error.mindee_http_error import MindeeHTTPError
 from mindee.product.invoice_splitter.invoice_splitter_v1 import InvoiceSplitterV1
 from mindee.product.receipt.receipt_v4 import ReceiptV4
 from tests.test_inputs import FILE_TYPES_DIR
@@ -41,7 +41,7 @@ def test_feedback_without_id(empty_client: Client):
 
 
 def test_parse_path_with_env_token(env_client: Client):
-    with pytest.raises(MindeeHTTPException):
+    with pytest.raises(MindeeHTTPError):
         input_doc = env_client.source_from_path(FILE_TYPES_DIR / "pdf" / "blank.pdf")
         env_client.parse(product.ReceiptV4, input_doc)
 
@@ -52,7 +52,7 @@ def test_parse_path_with_wrong_filetype(dummy_client: Client):
 
 
 def test_parse_path_with_wrong_token(dummy_client: Client):
-    with pytest.raises(MindeeHTTPException):
+    with pytest.raises(MindeeHTTPError):
         input_doc = dummy_client.source_from_path(FILE_TYPES_DIR / "pdf" / "blank.pdf")
         dummy_client.parse(product.ReceiptV4, input_doc)
 
@@ -70,7 +70,7 @@ def test_interface_version(dummy_client: Client):
         account_name="dummy",
         version="1.1",
     )
-    with pytest.raises(MindeeHTTPException):
+    with pytest.raises(MindeeHTTPError):
         input_doc = dummy_client.source_from_path(FILE_TYPES_DIR / "receipt.jpg")
         dummy_client.parse(product.CustomV1, input_doc, endpoint=dummy_endpoint)
 
@@ -81,7 +81,7 @@ def test_keep_file_open(dummy_client: Client):
     )
     try:
         dummy_client.parse(product.ReceiptV4, input_doc, close_file=False)
-    except MindeeHTTPException:
+    except MindeeHTTPError:
         pass
     assert not input_doc.file_object.closed
     input_doc.close()
@@ -100,7 +100,7 @@ def test_cut_options(dummy_client: Client):
             close_file=False,
             page_options=PageOptions(page_indexes=range(5)),
         )
-    except MindeeHTTPException:
+    except MindeeHTTPError:
         pass
     assert input_doc.count_doc_pages() == 5
     input_doc.close()
